Address PR #57 review feedback
- Change confidence from f32 to u8 (0-100 percentage) to preserve Eq trait
- Restore Eq derive on Resolution, HunkState, ConflictHunk, ParsedConflict
- Fix LocalProvider to use ProviderNotAvailable instead of ApiKeyError
- Fix AiStrategy to treat None confidence as below threshold when min_confidence > 0
- Add extract_json() to ClaudeProvider to handle code fences in responses
- Add with_timeout() constructor to ClaudeProvider for custom timeouts
- Fix clippy warnings (map_or, cast attributes)

+++ crates/weavr-ai/src/request.rs
@@ -34,8 +34,8 @@ pub struct AiRequest {
 pub struct AiResponse {
     /// The suggested merged content.
     pub suggestion: String,
-    /// Confidence score (0.0 to 1.0).
-    pub confidence: f32,
+    /// Confidence score (0-100 percentage).
+    pub confidence: u8,
     /// Explanation of the merge reasoning.
     pub explanation: Option<String>,
 }
